requirements/base.txt
Django==2.1.4
django-localflavor==2.1
psycopg2-binary==2.7.6.1
